[table-detection] - run drawings line merge only on table candidates
