+++ server/src/processing/TableDetectionModule/DrawingToTableHelper.ts
@@ -1,9 +1,25 @@
+/**
+ * Copyright 2020 AXA Group Operations S.A.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 import { Drawing, BoundingBox } from './../../types/DocumentRepresentation';
 import { JsonTable, JsonTableCell } from './TableDetectionModule';
 import { SvgLine } from '../../types/DocumentRepresentation/SvgLine';
 
 // when calculating rows, the minimum separation between lines for them to be considered as a row/col
-const MIN_ROW_HEIGHT = 10;
+const MIN_ROW_HEIGHT = 8;
 const MIN_COL_WIDTH = 10;
 
 export default (drawing: Drawing, pageHeight: number): JsonTable => {
@@ -37,10 +53,12 @@ function getTableWidth(cols: number[][]): number {
 }
 
 function getColsData(d: Drawing): number[][] {
+  const vBoxLines = d.box.toSvgLines().filter(l => l instanceof SvgLine && l.isVertical());
   const xPositions = d.content
+  .concat(vBoxLines)
     .filter(c => c instanceof SvgLine && c.isVertical())
     .map((l: SvgLine) => l.fromX)
-    .sort()
+    .sort((a, b) => a - b)
     .reduce(byMinimum(MIN_COL_WIDTH), []);
 
   const cols = [];
